Removed the documentation feature in favor of using the tests as documentation
